Reorganize core model crate
