+++ nautilus_core/model/src/ffi/data/mod.rs
@@ -13,24 +13,9 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use std::ffi::c_char;
-
-use nautilus_core::{ffi::string::str_to_cstr, time::UnixNanos};
-
-use super::ticker::Ticker;
-use crate::identifiers::instrument_id::InstrumentId;
-
-#[no_mangle]
-pub extern "C" fn ticker_new(
-    instrument_id: InstrumentId,
-    ts_event: UnixNanos,
-    ts_init: UnixNanos,
-) -> Ticker {
-    Ticker::new(instrument_id, ts_event, ts_init)
-}
-
-/// Returns a [`Ticker`] as a C string pointer.
-#[no_mangle]
-pub extern "C" fn ticker_to_cstr(ticker: &Ticker) -> *const c_char {
-    str_to_cstr(&ticker.to_string())
-}
+pub mod bar;
+pub mod delta;
+pub mod order;
+pub mod quote;
+pub mod ticker;
+pub mod trade;
