+++ nautilus_core/model/src/python/data/delta.rs
@@ -16,7 +16,6 @@
 use std::{
     collections::{hash_map::DefaultHasher, HashMap},
     hash::{Hash, Hasher},
-    str::FromStr,
 };
 
 use nautilus_core::{
@@ -24,108 +23,38 @@ use nautilus_core::{
     serialization::Serializable,
     time::UnixNanos,
 };
-use pyo3::{
-    prelude::*,
-    pyclass::CompareOp,
-    types::{PyDict, PyLong, PyString, PyTuple},
-};
+use pyo3::{prelude::*, pyclass::CompareOp, types::PyDict};
 
-use super::trade::TradeTick;
 use crate::{
-    enums::{AggressorSide, FromU8},
-    identifiers::{instrument_id::InstrumentId, trade_id::TradeId},
+    data::{delta::OrderBookDelta, order::BookOrder},
+    enums::BookAction,
+    identifiers::instrument_id::InstrumentId,
     python::PY_MODULE_MODEL,
-    types::{price::Price, quantity::Quantity},
 };
 
 #[pymethods]
-impl TradeTick {
+impl OrderBookDelta {
     #[new]
     fn py_new(
         instrument_id: InstrumentId,
-        price: Price,
-        size: Quantity,
-        aggressor_side: AggressorSide,
-        trade_id: TradeId,
+        action: BookAction,
+        order: BookOrder,
+        flags: u8,
+        sequence: u64,
         ts_event: UnixNanos,
         ts_init: UnixNanos,
     ) -> Self {
         Self::new(
             instrument_id,
-            price,
-            size,
-            aggressor_side,
-            trade_id,
+            action,
+            order,
+            flags,
+            sequence,
             ts_event,
             ts_init,
         )
     }
 
-    fn __setstate__(&mut self, py: Python, state: PyObject) -> PyResult<()> {
-        let tuple: (
-            &PyString,
-            &PyLong,
-            &PyLong,
-            &PyLong,
-            &PyLong,
-            &PyLong,
-            &PyString,
-            &PyLong,
-            &PyLong,
-        ) = state.extract(py)?;
-        let instrument_id_str: &str = tuple.0.extract()?;
-        let price_raw = tuple.1.extract()?;
-        let price_prec = tuple.2.extract()?;
-        let size_raw = tuple.3.extract()?;
-        let size_prec = tuple.4.extract()?;
-        let aggressor_side_u8 = tuple.5.extract()?;
-        let trade_id_str = tuple.6.extract()?;
-
-        self.instrument_id = InstrumentId::from_str(instrument_id_str).map_err(to_pyvalue_err)?;
-        self.price = Price::from_raw(price_raw, price_prec).map_err(to_pyvalue_err)?;
-        self.size = Quantity::from_raw(size_raw, size_prec).map_err(to_pyvalue_err)?;
-        self.aggressor_side = AggressorSide::from_u8(aggressor_side_u8).unwrap();
-        self.trade_id = TradeId::from_str(trade_id_str).map_err(to_pyvalue_err)?;
-        self.ts_event = tuple.7.extract()?;
-        self.ts_init = tuple.8.extract()?;
-
-        Ok(())
-    }
-
-    fn __getstate__(&self, _py: Python) -> PyResult<PyObject> {
-        Ok((
-            self.instrument_id.to_string(),
-            self.price.raw,
-            self.price.precision,
-            self.size.raw,
-            self.size.precision,
-            self.aggressor_side as u8,
-            self.trade_id.to_string(),
-            self.ts_event,
-            self.ts_init,
-        )
-            .to_object(_py))
-    }
-
-    fn __reduce__(&self, py: Python) -> PyResult<PyObject> {
-        let safe_constructor = py.get_type::<Self>().getattr("_safe_constructor")?;
-        let state = self.__getstate__(py)?;
-        Ok((safe_constructor, PyTuple::empty(py), state).to_object(py))
-    }
-
-    #[staticmethod]
-    fn _safe_constructor() -> PyResult<Self> {
-        Ok(Self::new(
-            InstrumentId::from("NULL.NULL"),
-            Price::zero(0),
-            Quantity::zero(0),
-            AggressorSide::NoAggressor,
-            TradeId::from("NULL"),
-            0,
-            0,
-        ))
-        // Safe default
-    }
     fn __richcmp__(&self, other: &Self, op: CompareOp, py: Python<'_>) -> Py<PyAny> {
         match op {
             CompareOp::Eq => self.eq(other).into_py(py),
@@ -145,7 +74,7 @@ impl TradeTick {
     }
 
     fn __repr__(&self) -> String {
-        format!("{}({})", stringify!(TradeTick), self)
+        format!("{self:?}")
     }
 
     #[getter]
@@ -154,23 +83,23 @@ impl TradeTick {
     }
 
     #[getter]
-    fn price(&self) -> Price {
-        self.price
+    fn action(&self) -> BookAction {
+        self.action
     }
 
     #[getter]
-    fn size(&self) -> Quantity {
-        self.size
+    fn order(&self) -> BookOrder {
+        self.order
     }
 
     #[getter]
-    fn aggressor_side(&self) -> AggressorSide {
-        self.aggressor_side
+    fn flags(&self) -> u8 {
+        self.flags
     }
 
     #[getter]
-    fn trade_id(&self) -> TradeId {
-        self.trade_id
+    fn sequence(&self) -> u64 {
+        self.sequence
     }
 
     #[getter]
@@ -186,7 +115,7 @@ impl TradeTick {
     #[staticmethod]
     #[pyo3(name = "fully_qualified_name")]
     fn py_fully_qualified_name() -> String {
-        format!("{}:{}", PY_MODULE_MODEL, stringify!(TradeTick))
+        format!("{}:{}", PY_MODULE_MODEL, stringify!(OrderBookDelta))
     }
 
     /// Return a dictionary representation of the object.
@@ -248,14 +177,14 @@ impl TradeTick {
     /// Return JSON encoded bytes representation of the object.
     #[pyo3(name = "as_json")]
     fn py_as_json(&self, py: Python<'_>) -> Py<PyAny> {
-        // SAFETY: Unwrapping is safe when serializing a valid object
+        // Unwrapping is safe when serializing a valid object
         self.as_json_bytes().unwrap().into_py(py)
     }
 
     /// Return MsgPack encoded bytes representation of the object.
     #[pyo3(name = "as_msgpack")]
     fn py_as_msgpack(&self, py: Python<'_>) -> Py<PyAny> {
-        // SAFETY: Unwrapping is safe when serializing a valid object
+        // Unwrapping is safe when serializing a valid object
         self.as_msgpack_bytes().unwrap().into_py(py)
     }
 }
@@ -265,44 +194,44 @@ impl TradeTick {
 ////////////////////////////////////////////////////////////////////////////////
 #[cfg(test)]
 mod tests {
-    use pyo3::{IntoPy, Python};
     use rstest::rstest;
 
-    use crate::data::trade::{stubs::*, TradeTick};
+    use super::*;
+    use crate::data::delta::stubs::stub_delta;
 
     #[rstest]
-    fn test_as_dict(trade_tick_ethusdt_buyer: TradeTick) {
+    fn test_as_dict(stub_delta: OrderBookDelta) {
         pyo3::prepare_freethreaded_python();
-        let tick = trade_tick_ethusdt_buyer;
+        let delta = stub_delta;
 
         Python::with_gil(|py| {
-            let dict_string = tick.py_as_dict(py).unwrap().to_string();
-            let expected_string = r#"{'type': 'TradeTick', 'instrument_id': 'ETHUSDT-PERP.BINANCE', 'price': '10000.0000', 'size': '1.00000000', 'aggressor_side': 'BUYER', 'trade_id': '123456789', 'ts_event': 0, 'ts_init': 1}"#;
+            let dict_string = delta.py_as_dict(py).unwrap().to_string();
+            let expected_string = r#"{'type': 'OrderBookDelta', 'instrument_id': 'AAPL.NASDAQ', 'action': 'ADD', 'order': {'side': 'BUY', 'price': '100.00', 'size': '10', 'order_id': 123456}, 'flags': 0, 'sequence': 1, 'ts_event': 1, 'ts_init': 2}"#;
             assert_eq!(dict_string, expected_string);
         });
     }
 
     #[rstest]
-    fn test_from_dict(trade_tick_ethusdt_buyer: TradeTick) {
+    fn test_from_dict(stub_delta: OrderBookDelta) {
         pyo3::prepare_freethreaded_python();
-        let tick = trade_tick_ethusdt_buyer;
+        let delta = stub_delta;
 
         Python::with_gil(|py| {
-            let dict = tick.py_as_dict(py).unwrap();
-            let parsed = TradeTick::py_from_dict(py, dict).unwrap();
-            assert_eq!(parsed, tick);
+            let dict = delta.py_as_dict(py).unwrap();
+            let parsed = OrderBookDelta::py_from_dict(py, dict).unwrap();
+            assert_eq!(parsed, delta);
         });
     }
 
     #[rstest]
-    fn test_from_pyobject(trade_tick_ethusdt_buyer: TradeTick) {
+    fn test_from_pyobject(stub_delta: OrderBookDelta) {
         pyo3::prepare_freethreaded_python();
-        let tick = trade_tick_ethusdt_buyer;
+        let delta = stub_delta;
 
         Python::with_gil(|py| {
-            let tick_pyobject = tick.into_py(py);
-            let parsed_tick = TradeTick::from_pyobject(tick_pyobject.as_ref(py)).unwrap();
-            assert_eq!(parsed_tick, tick);
+            let delta_pyobject = delta.into_py(py);
+            let parsed_delta = OrderBookDelta::from_pyobject(delta_pyobject.as_ref(py)).unwrap();
+            assert_eq!(parsed_delta, delta);
         });
     }
 }
