+++ nautilus_trader/core/includes/model.h
@@ -956,24 +956,6 @@ typedef struct Data_t {
 } Data_t;
 
 /**
- * Represents a single quote tick in a financial market.
- */
-typedef struct Ticker {
-    /**
-     * The quotes instrument ID.
-     */
-    struct InstrumentId_t instrument_id;
-    /**
-     * The UNIX timestamp (nanoseconds) when the tick event occurred.
-     */
-    uint64_t ts_event;
-    /**
-     * The UNIX timestamp (nanoseconds) when the data object was initialized.
-     */
-    uint64_t ts_init;
-} Ticker;
-
-/**
  * Represents a valid account ID.
  *
  * Must be correctly formatted with two valid strings either side of a hyphen '-'.
@@ -1139,6 +1121,24 @@ typedef struct Level_API {
     struct Level *_0;
 } Level_API;
 
+/**
+ * Represents a single quote tick in a financial market.
+ */
+typedef struct Ticker {
+    /**
+     * The quotes instrument ID.
+     */
+    struct InstrumentId_t instrument_id;
+    /**
+     * The UNIX timestamp (nanoseconds) when the tick event occurred.
+     */
+    uint64_t ts_event;
+    /**
+     * The UNIX timestamp (nanoseconds) when the data object was initialized.
+     */
+    uint64_t ts_init;
+} Ticker;
+
 typedef struct OrderDenied_t {
     struct TraderId_t trader_id;
     struct StrategyId_t strategy_id;
@@ -1229,184 +1229,6 @@ typedef struct Money_t {
 #define ERROR_PRICE (Price_t){ .raw = INT64_MAX, .precision = 0 }
 
 struct Data_t data_clone(const struct Data_t *data);
-
-struct BarSpecification_t bar_specification_new(uintptr_t step,
-                                                uint8_t aggregation,
-                                                uint8_t price_type);
-
-/**
- * Returns a [`BarSpecification`] as a C string pointer.
- */
-const char *bar_specification_to_cstr(const struct BarSpecification_t *bar_spec);
-
-uint64_t bar_specification_hash(const struct BarSpecification_t *bar_spec);
-
-uint8_t bar_specification_eq(const struct BarSpecification_t *lhs,
-                             const struct BarSpecification_t *rhs);
-
-uint8_t bar_specification_lt(const struct BarSpecification_t *lhs,
-                             const struct BarSpecification_t *rhs);
-
-uint8_t bar_specification_le(const struct BarSpecification_t *lhs,
-                             const struct BarSpecification_t *rhs);
-
-uint8_t bar_specification_gt(const struct BarSpecification_t *lhs,
-                             const struct BarSpecification_t *rhs);
-
-uint8_t bar_specification_ge(const struct BarSpecification_t *lhs,
-                             const struct BarSpecification_t *rhs);
-
-struct BarType_t bar_type_new(struct InstrumentId_t instrument_id,
-                              struct BarSpecification_t spec,
-                              uint8_t aggregation_source);
-
-/**
- * Returns any [`BarType`] parsing error from the provided C string pointer.
- *
- * # Safety
- *
- * - Assumes `ptr` is a valid C string pointer.
- */
-const char *bar_type_check_parsing(const char *ptr);
-
-/**
- * Returns a [`BarType`] from a C string pointer.
- *
- * # Safety
- *
- * - Assumes `ptr` is a valid C string pointer.
- */
-struct BarType_t bar_type_from_cstr(const char *ptr);
-
-uint8_t bar_type_eq(const struct BarType_t *lhs, const struct BarType_t *rhs);
-
-uint8_t bar_type_lt(const struct BarType_t *lhs, const struct BarType_t *rhs);
-
-uint8_t bar_type_le(const struct BarType_t *lhs, const struct BarType_t *rhs);
-
-uint8_t bar_type_gt(const struct BarType_t *lhs, const struct BarType_t *rhs);
-
-uint8_t bar_type_ge(const struct BarType_t *lhs, const struct BarType_t *rhs);
-
-uint64_t bar_type_hash(const struct BarType_t *bar_type);
-
-/**
- * Returns a [`BarType`] as a C string pointer.
- */
-const char *bar_type_to_cstr(const struct BarType_t *bar_type);
-
-struct Bar_t bar_new(struct BarType_t bar_type,
-                     struct Price_t open,
-                     struct Price_t high,
-                     struct Price_t low,
-                     struct Price_t close,
-                     struct Quantity_t volume,
-                     uint64_t ts_event,
-                     uint64_t ts_init);
-
-struct Bar_t bar_new_from_raw(struct BarType_t bar_type,
-                              int64_t open,
-                              int64_t high,
-                              int64_t low,
-                              int64_t close,
-                              uint8_t price_prec,
-                              uint64_t volume,
-                              uint8_t size_prec,
-                              uint64_t ts_event,
-                              uint64_t ts_init);
-
-uint8_t bar_eq(const struct Bar_t *lhs, const struct Bar_t *rhs);
-
-uint64_t bar_hash(const struct Bar_t *bar);
-
-/**
- * Returns a [`Bar`] as a C string.
- */
-const char *bar_to_cstr(const struct Bar_t *bar);
-
-struct OrderBookDelta_t orderbook_delta_new(struct InstrumentId_t instrument_id,
-                                            enum BookAction action,
-                                            struct BookOrder_t order,
-                                            uint8_t flags,
-                                            uint64_t sequence,
-                                            uint64_t ts_event,
-                                            uint64_t ts_init);
-
-uint8_t orderbook_delta_eq(const struct OrderBookDelta_t *lhs, const struct OrderBookDelta_t *rhs);
-
-uint64_t orderbook_delta_hash(const struct OrderBookDelta_t *delta);
-
-struct BookOrder_t book_order_from_raw(enum OrderSide order_side,
-                                       int64_t price_raw,
-                                       uint8_t price_prec,
-                                       uint64_t size_raw,
-                                       uint8_t size_prec,
-                                       uint64_t order_id);
-
-uint8_t book_order_eq(const struct BookOrder_t *lhs, const struct BookOrder_t *rhs);
-
-uint64_t book_order_hash(const struct BookOrder_t *order);
-
-double book_order_exposure(const struct BookOrder_t *order);
-
-double book_order_signed_size(const struct BookOrder_t *order);
-
-/**
- * Returns a [`BookOrder`] display string as a C string pointer.
- */
-const char *book_order_display_to_cstr(const struct BookOrder_t *order);
-
-/**
- * Returns a [`BookOrder`] debug string as a C string pointer.
- */
-const char *book_order_debug_to_cstr(const struct BookOrder_t *order);
-
-struct QuoteTick_t quote_tick_new(struct InstrumentId_t instrument_id,
-                                  int64_t bid_price_raw,
-                                  int64_t ask_price_raw,
-                                  uint8_t bid_price_prec,
-                                  uint8_t ask_price_prec,
-                                  uint64_t bid_size_raw,
-                                  uint64_t ask_size_raw,
-                                  uint8_t bid_size_prec,
-                                  uint8_t ask_size_prec,
-                                  uint64_t ts_event,
-                                  uint64_t ts_init);
-
-uint8_t quote_tick_eq(const struct QuoteTick_t *lhs, const struct QuoteTick_t *rhs);
-
-uint64_t quote_tick_hash(const struct QuoteTick_t *delta);
-
-/**
- * Returns a [`QuoteTick`] as a C string pointer.
- */
-const char *quote_tick_to_cstr(const struct QuoteTick_t *tick);
-
-struct Ticker ticker_new(struct InstrumentId_t instrument_id, uint64_t ts_event, uint64_t ts_init);
-
-/**
- * Returns a [`Ticker`] as a C string pointer.
- */
-const char *ticker_to_cstr(const struct Ticker *ticker);
-
-struct TradeTick_t trade_tick_new(struct InstrumentId_t instrument_id,
-                                  int64_t price_raw,
-                                  uint8_t price_prec,
-                                  uint64_t size_raw,
-                                  uint8_t size_prec,
-                                  enum AggressorSide aggressor_side,
-                                  struct TradeId_t trade_id,
-                                  uint64_t ts_event,
-                                  uint64_t ts_init);
-
-uint8_t trade_tick_eq(const struct TradeTick_t *lhs, const struct TradeTick_t *rhs);
-
-uint64_t trade_tick_hash(const struct TradeTick_t *delta);
-
-/**
- * Returns a [`TradeTick`] as a C string pointer.
- */
-const char *trade_tick_to_cstr(const struct TradeTick_t *tick);
 
 const char *account_type_to_cstr(enum AccountType value);
 
@@ -1983,6 +1805,184 @@ double level_exposure(const struct Level_API *level);
 void vec_levels_drop(CVec v);
 
 void vec_orders_drop(CVec v);
+
+struct BarSpecification_t bar_specification_new(uintptr_t step,
+                                                uint8_t aggregation,
+                                                uint8_t price_type);
+
+/**
+ * Returns a [`BarSpecification`] as a C string pointer.
+ */
+const char *bar_specification_to_cstr(const struct BarSpecification_t *bar_spec);
+
+uint64_t bar_specification_hash(const struct BarSpecification_t *bar_spec);
+
+uint8_t bar_specification_eq(const struct BarSpecification_t *lhs,
+                             const struct BarSpecification_t *rhs);
+
+uint8_t bar_specification_lt(const struct BarSpecification_t *lhs,
+                             const struct BarSpecification_t *rhs);
+
+uint8_t bar_specification_le(const struct BarSpecification_t *lhs,
+                             const struct BarSpecification_t *rhs);
+
+uint8_t bar_specification_gt(const struct BarSpecification_t *lhs,
+                             const struct BarSpecification_t *rhs);
+
+uint8_t bar_specification_ge(const struct BarSpecification_t *lhs,
+                             const struct BarSpecification_t *rhs);
+
+struct BarType_t bar_type_new(struct InstrumentId_t instrument_id,
+                              struct BarSpecification_t spec,
+                              uint8_t aggregation_source);
+
+/**
+ * Returns any [`BarType`] parsing error from the provided C string pointer.
+ *
+ * # Safety
+ *
+ * - Assumes `ptr` is a valid C string pointer.
+ */
+const char *bar_type_check_parsing(const char *ptr);
+
+/**
+ * Returns a [`BarType`] from a C string pointer.
+ *
+ * # Safety
+ *
+ * - Assumes `ptr` is a valid C string pointer.
+ */
+struct BarType_t bar_type_from_cstr(const char *ptr);
+
+uint8_t bar_type_eq(const struct BarType_t *lhs, const struct BarType_t *rhs);
+
+uint8_t bar_type_lt(const struct BarType_t *lhs, const struct BarType_t *rhs);
+
+uint8_t bar_type_le(const struct BarType_t *lhs, const struct BarType_t *rhs);
+
+uint8_t bar_type_gt(const struct BarType_t *lhs, const struct BarType_t *rhs);
+
+uint8_t bar_type_ge(const struct BarType_t *lhs, const struct BarType_t *rhs);
+
+uint64_t bar_type_hash(const struct BarType_t *bar_type);
+
+/**
+ * Returns a [`BarType`] as a C string pointer.
+ */
+const char *bar_type_to_cstr(const struct BarType_t *bar_type);
+
+struct Bar_t bar_new(struct BarType_t bar_type,
+                     struct Price_t open,
+                     struct Price_t high,
+                     struct Price_t low,
+                     struct Price_t close,
+                     struct Quantity_t volume,
+                     uint64_t ts_event,
+                     uint64_t ts_init);
+
+struct Bar_t bar_new_from_raw(struct BarType_t bar_type,
+                              int64_t open,
+                              int64_t high,
+                              int64_t low,
+                              int64_t close,
+                              uint8_t price_prec,
+                              uint64_t volume,
+                              uint8_t size_prec,
+                              uint64_t ts_event,
+                              uint64_t ts_init);
+
+uint8_t bar_eq(const struct Bar_t *lhs, const struct Bar_t *rhs);
+
+uint64_t bar_hash(const struct Bar_t *bar);
+
+/**
+ * Returns a [`Bar`] as a C string.
+ */
+const char *bar_to_cstr(const struct Bar_t *bar);
+
+struct OrderBookDelta_t orderbook_delta_new(struct InstrumentId_t instrument_id,
+                                            enum BookAction action,
+                                            struct BookOrder_t order,
+                                            uint8_t flags,
+                                            uint64_t sequence,
+                                            uint64_t ts_event,
+                                            uint64_t ts_init);
+
+uint8_t orderbook_delta_eq(const struct OrderBookDelta_t *lhs, const struct OrderBookDelta_t *rhs);
+
+uint64_t orderbook_delta_hash(const struct OrderBookDelta_t *delta);
+
+struct BookOrder_t book_order_from_raw(enum OrderSide order_side,
+                                       int64_t price_raw,
+                                       uint8_t price_prec,
+                                       uint64_t size_raw,
+                                       uint8_t size_prec,
+                                       uint64_t order_id);
+
+uint8_t book_order_eq(const struct BookOrder_t *lhs, const struct BookOrder_t *rhs);
+
+uint64_t book_order_hash(const struct BookOrder_t *order);
+
+double book_order_exposure(const struct BookOrder_t *order);
+
+double book_order_signed_size(const struct BookOrder_t *order);
+
+/**
+ * Returns a [`BookOrder`] display string as a C string pointer.
+ */
+const char *book_order_display_to_cstr(const struct BookOrder_t *order);
+
+/**
+ * Returns a [`BookOrder`] debug string as a C string pointer.
+ */
+const char *book_order_debug_to_cstr(const struct BookOrder_t *order);
+
+struct QuoteTick_t quote_tick_new(struct InstrumentId_t instrument_id,
+                                  int64_t bid_price_raw,
+                                  int64_t ask_price_raw,
+                                  uint8_t bid_price_prec,
+                                  uint8_t ask_price_prec,
+                                  uint64_t bid_size_raw,
+                                  uint64_t ask_size_raw,
+                                  uint8_t bid_size_prec,
+                                  uint8_t ask_size_prec,
+                                  uint64_t ts_event,
+                                  uint64_t ts_init);
+
+uint8_t quote_tick_eq(const struct QuoteTick_t *lhs, const struct QuoteTick_t *rhs);
+
+uint64_t quote_tick_hash(const struct QuoteTick_t *delta);
+
+/**
+ * Returns a [`QuoteTick`] as a C string pointer.
+ */
+const char *quote_tick_to_cstr(const struct QuoteTick_t *tick);
+
+struct Ticker ticker_new(struct InstrumentId_t instrument_id, uint64_t ts_event, uint64_t ts_init);
+
+/**
+ * Returns a [`Ticker`] as a C string pointer.
+ */
+const char *ticker_to_cstr(const struct Ticker *ticker);
+
+struct TradeTick_t trade_tick_new(struct InstrumentId_t instrument_id,
+                                  int64_t price_raw,
+                                  uint8_t price_prec,
+                                  uint64_t size_raw,
+                                  uint8_t size_prec,
+                                  enum AggressorSide aggressor_side,
+                                  struct TradeId_t trade_id,
+                                  uint64_t ts_event,
+                                  uint64_t ts_init);
+
+uint8_t trade_tick_eq(const struct TradeTick_t *lhs, const struct TradeTick_t *rhs);
+
+uint64_t trade_tick_hash(const struct TradeTick_t *delta);
+
+/**
+ * Returns a [`TradeTick`] as a C string pointer.
+ */
+const char *trade_tick_to_cstr(const struct TradeTick_t *tick);
 
 /**
  * # Safety
